feat: 커스텀 예외 추가, 카운트 쿼리 작성

+++ src/main/java/com/deokhugam/domain/poweruser/dto/response/PowerUserDto.java
@@ -20,7 +20,7 @@ public class PowerUserDto {
     private Double score;
     private Long likeCount;
     private Long commentCount;
-    private Long reviewScoreSum;
+    private Double reviewScoreSum;
 
     public static PowerUserDto from(PowerUser powerUser) {
         return new PowerUserDto(
